refactor: move Ethereum-specific helpers to the crate
