+++ chain-signatures/chain-ethereum/src/lib.rs
@@ -6,9 +6,10 @@ pub mod indexer_eth_direct_rpc;
 #[cfg(feature = "helios")]
 pub mod indexer_eth_helios;
 pub mod publisher;
-pub mod util;
+pub mod respond_bidirectional;
 #[cfg(test)]
 pub mod test_utils;
+pub mod util;
 
 pub use config::EthConfig;
 pub use indexer::EthereumStream;
