+++ chain-signatures/chain-ethereum/src/respond_bidirectional.rs
@@ -1,0 +1,281 @@
+use alloy::primitives::{Bytes, I256, U256};
+use alloy_dyn_abi::{DynSolType, DynSolValue};
+use borsh::BorshSerialize;
+use mpc_primitives::SerDeserFormat;
+use serde_json::Value;
+use std::collections::HashMap;
+use std::io::Write;
+
+// Use Abi as this is what we are using for ethereum
+const OUTPUT_DESERIALIZATION_FORMAT: SerDeserFormat = SerDeserFormat::Abi;
+
+#[derive(Debug, Clone, PartialEq, Eq, Hash, serde::Serialize, serde::Deserialize)]
+struct AbiField {
+    name: String,
+    #[serde(rename = "type")]
+    typ: String,
+}
+
+#[derive(Debug, Clone, Default)]
+pub struct Output {
+    fields: HashMap<String, DynSolValue>,
+    /// `true` when this `Output` was built from a real ETH contract-call return
+    /// (via `TransactionOutput::from_call_result`); `false` for the
+    /// `non_contract_call_output()` path (plain transfers). Drives whether
+    /// `serialize` encodes real data or synthesizes per-schema defaults.
+    from_contract_call: bool,
+}
+
+impl Output {
+    pub fn is_contract_call(&self) -> bool {
+        self.from_contract_call
+    }
+
+    /// Encode this output for the given format using `schema_json_bytes` as
+    /// the field shape. For non-contract-call outputs (plain transfers),
+    /// synthesizes per-field default values from the schema. Real decoded
+    /// data from `from_call_result` flows through unchanged.
+    pub fn serialize(
+        &self,
+        format: SerDeserFormat,
+        schema_json_bytes: &[u8],
+    ) -> anyhow::Result<Vec<u8>> {
+        let schema = parse_schema_fields(schema_json_bytes)?;
+        let data_owned;
+        let data = if self.is_contract_call() {
+            self
+        } else {
+            data_owned = default_output_for_non_contract_call(&schema)?;
+            &data_owned
+        };
+        match format {
+            SerDeserFormat::Abi => encode_abi(data, &schema),
+            SerDeserFormat::Borsh => encode_borsh(data, &schema),
+        }
+    }
+}
+
+#[derive(Debug)]
+pub struct TransactionOutput {
+    pub success: bool,
+    pub output: Output,
+}
+
+impl TransactionOutput {
+    pub fn non_contract_call_output() -> Self {
+        Self {
+            success: true,
+            output: Output {
+                fields: HashMap::new(),
+                from_contract_call: false,
+            },
+        }
+    }
+
+    pub fn from_call_result(schema_json: &[u8], call_result: &Bytes) -> anyhow::Result<Self> {
+        let schema: Vec<AbiField> = serde_json::from_slice(schema_json)
+            .map_err(|e| anyhow::anyhow!("Failed to get abi fields from schema: {e:?}"))?;
+
+        let types: Vec<DynSolType> = schema
+            .iter()
+            .map(|f| f.typ.parse())
+            .collect::<Result<_, _>>()
+            .map_err(|e| anyhow::anyhow!("Failed to parse eth transaction types: {e:?}"))?;
+
+        let tuple_type = DynSolType::Tuple(types);
+
+        let DynSolValue::Tuple(values) = tuple_type
+            .abi_decode(call_result)
+            .map_err(|e| anyhow::anyhow!("Failed to tuple types: {e:?}"))?
+        else {
+            anyhow::bail!("Can't decode to tuple type");
+        };
+
+        let mut output_map = HashMap::new();
+        for (field, value) in schema.into_iter().zip(values) {
+            output_map.insert(field.name, value);
+        }
+
+        Ok(TransactionOutput {
+            success: true,
+            output: Output {
+                fields: output_map,
+                from_contract_call: true,
+            },
+        })
+    }
+}
+
+/// Decode a transaction's output and re-serialize it for the respond chain.
+///
+/// `trace_output` is the `debug_traceTransaction` return data, required when
+/// `is_contract_call` is true.
+pub fn build_serialized_output(
+    is_contract_call: bool,
+    output_deserialization_schema: &[u8],
+    trace_output: Option<&Bytes>,
+    respond_serialization_format: SerDeserFormat,
+    respond_serialization_schema: &[u8],
+) -> anyhow::Result<Vec<u8>> {
+    let transaction_output = match OUTPUT_DESERIALIZATION_FORMAT {
+        SerDeserFormat::Abi if is_contract_call => {
+            let trace_output = trace_output.ok_or_else(|| {
+                anyhow::anyhow!("contract-call output extraction requires trace output")
+            })?;
+            TransactionOutput::from_call_result(output_deserialization_schema, trace_output)?
+        }
+        _ => TransactionOutput::non_contract_call_output(),
+    };
+
+    transaction_output
+        .output
+        .serialize(respond_serialization_format, respond_serialization_schema)
+}
+
+fn encode_abi(data: &Output, schema: &[AbiField]) -> anyhow::Result<Vec<u8>> {
+    let values = schema
+        .iter()
+        .map(|field| match data.fields.get(&field.name) {
+            Some(value) => Ok(value.clone()),
+            None => Err(anyhow::anyhow!(
+                "Missing required field '{}' in output",
+                field.name
+            )),
+        })
+        .collect::<Result<Vec<_>, _>>()?;
+    encode_abi_values(schema, &values)
+}
+
+fn encode_borsh(data: &Output, schema: &[AbiField]) -> anyhow::Result<Vec<u8>> {
+    assert!(
+        schema.len() == 1,
+        "borsh schema must have exactly one field"
+    );
+    let val = data
+        .fields
+        .get(&schema[0].name)
+        .ok_or_else(|| anyhow::anyhow!("missing value for field '{}'", schema[0].name))?;
+    let mut buf = Vec::with_capacity(128);
+    serialize_dynsol(&mut buf, val)?;
+    Ok(buf)
+}
+
+fn encode_abi_values(schema: &[AbiField], values: &[DynSolValue]) -> anyhow::Result<Vec<u8>> {
+    if schema.len() != values.len() {
+        anyhow::bail!(
+            "Schema and values length mismatch: {} != {}",
+            schema.len(),
+            values.len()
+        );
+    }
+    for (f, v) in schema.iter().zip(values.iter()) {
+        let ty: DynSolType = f.typ.parse()?;
+        if !ty.matches(v) {
+            anyhow::bail!("Value {v:?} doesn't match Solidity type {}", f.typ);
+        }
+    }
+    let mut combined = Vec::new();
+    for v in values {
+        combined.extend(v.abi_encode());
+    }
+    Ok(combined)
+}
+
+fn serialize_dynsol<W: Write>(w: &mut W, v: &DynSolValue) -> anyhow::Result<()> {
+    use DynSolValue::*;
+    match v {
+        Bool(b) => {
+            b.serialize(w)?;
+        }
+        Address(a) => a.serialize(w)?,
+        Uint(u, size) => write_u256(w, *u, *size)?,
+        Int(i, size) => write_i256(w, *i, *size)?,
+        FixedBytes(b, _) => w.write_all(b.as_slice())?,
+        Bytes(b) => b.serialize(w)?,
+        String(s) => s.serialize(w)?,
+        Array(xs) => {
+            (xs.len() as u32).serialize(w)?;
+            for x in xs {
+                serialize_dynsol(w, x)?;
+            }
+        }
+        FixedArray(xs) => {
+            for x in xs {
+                serialize_dynsol(w, x)?;
+            }
+        }
+        Tuple(xs) => {
+            for x in xs {
+                serialize_dynsol(w, x)?;
+            }
+        }
+        other => anyhow::bail!("unsupported DynSolValue variant: {other:?}"),
+    }
+    Ok(())
+}
+
+fn write_u256<W: Write>(w: &mut W, x: U256, size: usize) -> anyhow::Result<()> {
+    let le = x.to_le_bytes::<{ U256::BYTES }>();
+    w.write_all(&le[..size.min(U256::BYTES)])
+        .map_err(Into::into)
+}
+
+fn write_i256<W: Write>(w: &mut W, x: I256, size: usize) -> anyhow::Result<()> {
+    let le = x.to_le_bytes::<{ I256::BYTES }>();
+    w.write_all(&le[..size.min(I256::BYTES)])
+        .map_err(Into::into)
+}
+
+/// Parse a schema JSON describing the response shape. Accepts a JSON array of
+/// `{name, type}` objects (canonical form), a single object (treated as a
+/// one-field schema), or a bare string (treated as a single typed field with
+/// an empty name).
+fn parse_schema_fields(schema_json_bytes: &[u8]) -> anyhow::Result<Vec<AbiField>> {
+    let v: Value = serde_json::from_slice(schema_json_bytes)
+        .map_err(|e| anyhow::anyhow!("schema JSON parse failed: {e:?}"))?;
+
+    Ok(match v {
+        Value::Array(arr) => arr
+            .into_iter()
+            .map(|item| {
+                serde_json::from_value(item)
+                    .map_err(|e| anyhow::anyhow!("invalid field in array: {e:?}"))
+            })
+            .collect::<Result<Vec<_>, anyhow::Error>>()?,
+        Value::Object(obj) => {
+            vec![serde_json::from_value(Value::Object(obj))
+                .map_err(|e| anyhow::anyhow!("invalid single object schema: {e:?}"))?]
+        }
+        Value::String(s) => vec![AbiField {
+            name: String::new(),
+            typ: s,
+        }],
+        other => anyhow::bail!("unsupported schema JSON shape: {other}"),
+    })
+}
+
+/// Synthesize per-field default values for an `Output` whose source tx was
+/// not a contract function call.
+fn default_output_for_non_contract_call(schema: &[AbiField]) -> anyhow::Result<Output> {
+    let mut data = HashMap::new();
+    for field in schema {
+        match field.typ.as_str() {
+            "string" => {
+                data.insert(
+                    field.name.clone(),
+                    DynSolValue::String("non_function_call_success".to_string()),
+                );
+            }
+            "bool" => {
+                data.insert(field.name.clone(), DynSolValue::Bool(true));
+            }
+            other => anyhow::bail!(
+                "cannot synthesize default for non-function-call output of type {other}"
+            ),
+        }
+    }
+    Ok(Output {
+        fields: data,
+        from_contract_call: false,
+    })
+}
