+++ chain-signatures/node/src/lib.rs
@@ -6,7 +6,6 @@ pub mod config;
 pub mod gcp;
 pub mod indexer;
 
-pub mod indexer_eth;
 pub mod indexer_hydration;
 pub mod indexer_sol;
 pub mod kdf;
