+++ chain-signatures/node/src/sign_bidirectional.rs
@@ -1,29 +1,16 @@
 use crate::protocol::{Chain, IndexedSignRequest};
-use alloy::primitives::{keccak256, Address, Bytes, I256, U256};
-use alloy_dyn_abi::{DynSolType, DynSolValue};
-use borsh::BorshSerialize;
+use alloy::primitives::{keccak256, Address, Bytes};
 use cait_sith::protocol::Participant;
 use k256::elliptic_curve::point::AffineCoordinates;
 use k256::elliptic_curve::sec1::ToEncodedPoint as _;
 use k256::{AffinePoint, Scalar};
 use mpc_crypto::derive_key;
 use mpc_primitives::{
-    BidirectionalTx, ChainFromError, SerDeserFormat, SignBidirectionalEvent, Signature,
+    BidirectionalTx, ChainFromError, SignBidirectionalEvent, Signature,
 };
 use rlp::{Rlp, RlpStream};
-use serde_json::Value;
-
-use std::collections::HashMap;
-use std::io::Write;
 
 pub type RequestId = [u8; 32];
-
-#[derive(Debug, Clone, PartialEq, Eq, Hash, serde::Serialize, serde::Deserialize)]
-struct AbiField {
-    name: String,
-    #[serde(rename = "type")]
-    typ: String,
-}
 
 #[derive(Debug, Clone, PartialEq, Eq, serde::Serialize, serde::Deserialize)]
 pub struct PublishState {
@@ -163,127 +150,6 @@ impl BidirectionalTxExt for BidirectionalTx {
             )),
             _ => anyhow::bail!("Unsupported chain: {}", self.source_chain),
         }
-    }
-}
-
-#[derive(Debug, Clone, Default)]
-pub struct Output {
-    fields: HashMap<String, DynSolValue>,
-    /// `true` when this `Output` was built from a real ETH contract-call return
-    /// (via `TransactionOutput::from_call_result`); `false` for the
-    /// `non_contract_call_output()` path (plain transfers). Drives whether
-    /// `serialize` encodes real data or synthesizes per-schema defaults.
-    from_contract_call: bool,
-}
-
-impl Output {
-    pub fn is_contract_call(&self) -> bool {
-        self.from_contract_call
-    }
-
-    /// Encode this output for the given format using `schema_json_bytes` as
-    /// the field shape. For non-contract-call outputs (plain transfers),
-    /// synthesizes per-field default values from the schema. Real decoded
-    /// data from `from_call_result` flows through unchanged.
-    pub fn serialize(
-        &self,
-        format: SerDeserFormat,
-        schema_json_bytes: &[u8],
-    ) -> anyhow::Result<Vec<u8>> {
-        let schema = parse_schema_fields(schema_json_bytes)?;
-        let data_owned;
-        let data = if self.is_contract_call() {
-            self
-        } else {
-            data_owned = default_output_for_non_contract_call(&schema)?;
-            &data_owned
-        };
-        match format {
-            SerDeserFormat::Abi => encode_abi(data, &schema),
-            SerDeserFormat::Borsh => encode_borsh(data, &schema),
-        }
-    }
-}
-
-fn encode_abi(data: &Output, schema: &[AbiField]) -> anyhow::Result<Vec<u8>> {
-    let values = schema
-        .iter()
-        .map(|field| match data.fields.get(&field.name) {
-            Some(value) => Ok(value.clone()),
-            None => Err(anyhow::anyhow!(
-                "Missing required field '{}' in output",
-                field.name
-            )),
-        })
-        .collect::<Result<Vec<_>, _>>()?;
-    encode_abi_values(schema, &values)
-}
-
-fn encode_borsh(data: &Output, schema: &[AbiField]) -> anyhow::Result<Vec<u8>> {
-    assert!(
-        schema.len() == 1,
-        "borsh schema must have exactly one field"
-    );
-    let val = data
-        .fields
-        .get(&schema[0].name)
-        .ok_or_else(|| anyhow::anyhow!("missing value for field '{}'", schema[0].name))?;
-    let mut buf = Vec::with_capacity(128);
-    serialize_dynsol(&mut buf, val)?;
-    Ok(buf)
-}
-
-#[derive(Debug)]
-pub struct TransactionOutput {
-    pub success: bool,
-    pub output: Output,
-}
-
-impl TransactionOutput {
-    pub fn non_contract_call_output() -> Self {
-        Self {
-            success: true,
-            output: Output {
-                fields: HashMap::new(),
-                from_contract_call: false,
-            },
-        }
-    }
-
-    pub fn from_call_result(schema_json: &[u8], call_result: &Bytes) -> anyhow::Result<Self> {
-        let schema: Vec<AbiField> = serde_json::from_slice(schema_json)
-            .map_err(|e| anyhow::anyhow!("Failed to get abi fields from schema: {e:?}"))?;
-
-        let types: Vec<DynSolType> = schema
-            .iter()
-            .map(|f| f.typ.parse()) // calls DynSolType::parse via FromStr
-            .collect::<Result<_, _>>()
-            .map_err(|e| anyhow::anyhow!("Failed to parse eth transaction types: {e:?}"))?;
-
-        // Build a single tuple DynSolType
-        let tuple_type = DynSolType::Tuple(types);
-
-        // Decode the whole result as a tuple
-        let DynSolValue::Tuple(values) = tuple_type
-            .abi_decode(call_result)
-            .map_err(|e| anyhow::anyhow!("Failed to tuple types: {e:?}"))?
-        else {
-            anyhow::bail!("Can't decode to tuple type");
-        };
-
-        // Map to named output
-        let mut output_map = HashMap::new();
-        for (field, value) in schema.into_iter().zip(values) {
-            output_map.insert(field.name, value);
-        }
-
-        Ok(TransactionOutput {
-            success: true,
-            output: Output {
-                fields: output_map,
-                from_contract_call: true,
-            },
-        })
     }
 }
 
@@ -468,34 +334,6 @@ pub fn derive_user_address(mpc_pk: mpc_crypto::PublicKey, derivation_epsilon: Sc
     public_key_to_address(user_pk.to_encoded_point(false).as_bytes())
 }
 
-/// Synthesize per-field default values for an `Output` whose source tx was
-/// not a contract function call. The destination chain's contract still needs
-/// shaped bytes back, so we fill defaults from the schema: `bool` → `true`,
-/// `string` → `"non_function_call_success"`. Other field types are unsupported.
-fn default_output_for_non_contract_call(schema: &[AbiField]) -> anyhow::Result<Output> {
-    let mut data = HashMap::new();
-    for field in schema {
-        match field.typ.as_str() {
-            "string" => {
-                data.insert(
-                    field.name.clone(),
-                    DynSolValue::String("non_function_call_success".to_string()),
-                );
-            }
-            "bool" => {
-                data.insert(field.name.clone(), DynSolValue::Bool(true));
-            }
-            other => anyhow::bail!(
-                "cannot synthesize default for non-function-call output of type {other}"
-            ),
-        }
-    }
-    Ok(Output {
-        fields: data,
-        from_contract_call: false,
-    })
-}
-
 #[cfg(test)]
 mod derive_tests {
     use super::derive_user_address;
@@ -521,125 +359,6 @@ mod derive_tests {
 
         assert_eq!(derive_user_address(mpc_pk, derivation_epsilon), expected);
     }
-}
-
-fn encode_abi_values(schema: &[AbiField], values: &[DynSolValue]) -> anyhow::Result<Vec<u8>> {
-    if schema.len() != values.len() {
-        anyhow::bail!(
-            "Schema and values length mismatch: {} != {}",
-            schema.len(),
-            values.len()
-        );
-    }
-    for (f, v) in schema.iter().zip(values.iter()) {
-        let ty: DynSolType = f.typ.parse()?;
-        if !ty.matches(v) {
-            anyhow::bail!("Value {v:?} doesn't match Solidity type {}", f.typ);
-        }
-    }
-    // Encode each value and concatenate
-    let mut combined = Vec::new();
-    for v in values {
-        combined.extend(v.abi_encode());
-    }
-
-    Ok(combined)
-}
-
-/* ---------- DynSolValue -> Borsh serializer (runtime) ---------- */
-
-fn serialize_dynsol<W: Write>(w: &mut W, v: &DynSolValue) -> anyhow::Result<()> {
-    use DynSolValue::*;
-    match v {
-        // -------- Primitives --------
-        Bool(b) => {
-            // Borsh bool is u8 (0/1) via BorshSerialize on bool
-            b.serialize(w)?;
-        }
-        Address(a) => a.serialize(w)?,
-        Uint(u, size) => write_u256(w, *u, *size)?,
-        Int(i, size) => write_i256(w, *i, *size)?,
-
-        // -------- Bytes-like --------
-        // Fixed bytes -> raw bytes (no length)
-        FixedBytes(b, _) => w.write_all(b.as_slice())?,
-        // Dynamic bytes -> Vec<u8> (u32 length + bytes)
-        Bytes(b) => b.serialize(w)?,
-
-        // -------- Strings --------
-        String(s) => s.serialize(w)?,
-
-        // -------- Arrays --------
-        // Dynamic array -> Borsh Vec<T>: u32 length + elements
-        Array(xs) => {
-            (xs.len() as u32).serialize(w)?;
-            for x in xs {
-                serialize_dynsol(w, x)?;
-            }
-        }
-        // Fixed array -> elements inline (no length)
-        FixedArray(xs) => {
-            for x in xs {
-                serialize_dynsol(w, x)?;
-            }
-        }
-
-        // -------- Tuple --------
-        // Concatenate members
-        Tuple(xs) => {
-            for x in xs {
-                serialize_dynsol(w, x)?;
-            }
-        }
-
-        // Add more variants if you use them (e.g., custom types).
-        other => anyhow::bail!("unsupported DynSolValue variant: {other:?}"),
-    }
-    Ok(())
-}
-
-/* ------------------ helpers using borsh where possible ------------------ */
-
-fn write_u256<W: Write>(w: &mut W, x: U256, size: usize) -> anyhow::Result<()> {
-    // Use the size parameter to determine how many bytes to write
-    let le = x.to_le_bytes::<{ U256::BYTES }>();
-    w.write_all(&le[..size.min(U256::BYTES)])
-        .map_err(Into::into)
-}
-
-fn write_i256<W: Write>(w: &mut W, x: I256, size: usize) -> anyhow::Result<()> {
-    // Use the size parameter to determine how many bytes to write
-    let le = x.to_le_bytes::<{ I256::BYTES }>();
-    w.write_all(&le[..size.min(I256::BYTES)])
-        .map_err(Into::into)
-}
-
-/// Parse a schema JSON describing the response shape. Accepts a JSON array of
-/// `{name, type}` objects (canonical form), a single object (treated as a
-/// one-field schema), or a bare string (treated as a single typed field with
-/// an empty name).
-fn parse_schema_fields(schema_json_bytes: &[u8]) -> anyhow::Result<Vec<AbiField>> {
-    let v: Value = serde_json::from_slice(schema_json_bytes)
-        .map_err(|e| anyhow::anyhow!("schema JSON parse failed: {e:?}"))?;
-
-    Ok(match v {
-        Value::Array(arr) => arr
-            .into_iter()
-            .map(|item| {
-                serde_json::from_value(item)
-                    .map_err(|e| anyhow::anyhow!("invalid field in array: {e:?}"))
-            })
-            .collect::<Result<Vec<_>, anyhow::Error>>()?,
-        Value::Object(obj) => {
-            vec![serde_json::from_value(Value::Object(obj))
-                .map_err(|e| anyhow::anyhow!("invalid single object schema: {e:?}"))?]
-        }
-        Value::String(s) => vec![AbiField {
-            name: String::new(),
-            typ: s,
-        }],
-        other => anyhow::bail!("unsupported schema JSON shape: {other}"),
-    })
 }
 
 #[derive(Clone)]
