+++ integration-tests/src/actions/sign.rs
@@ -18,9 +18,9 @@ use elliptic_curve::sec1::FromEncodedPoint;
 use futures::StreamExt;
 use generic_array::GenericArray;
 use k256::Secp256k1;
+use mpc_chain_ethereum::abi::{ChainSignatures, SignatureRequestedEncoding};
 use mpc_contract::primitives::SignRequest;
 use mpc_crypto::ScalarExt as _;
-use mpc_node::indexer_eth::abi::{ChainSignatures, SignatureRequestedEncoding};
 use mpc_primitives::LATEST_MPC_KEY_VERSION;
 use near_crypto::InMemorySigner;
 use near_fetch::ops::AsyncTransactionStatus;
