+++ integration-tests/src/lib.rs
@@ -20,10 +20,10 @@ use alloy::primitives::{Address, U256};
 use anyhow::Context as _;
 use cluster::spawner::ClusterSpawner;
 use mpc_chain_canton::CantonConfig;
+use mpc_chain_ethereum::EthConfig;
 use mpc_contract::config::{PresignatureConfig, ProtocolConfig, TripleConfig};
 use mpc_contract::primitives::CandidateInfo;
 use mpc_node::gcp::GcpService;
-use mpc_node::indexer_eth::EthConfig;
 use mpc_node::indexer_hydration::HydrationConfig;
 use mpc_node::indexer_sol::SolConfig;
 use mpc_node::{logs, mesh, node_client, storage};
